[~] split dsidekiq classes into files
